feat(测试计划、接口场景): 修复场景批量编辑失败的缺陷、为测试计划的定时任务和Jenkins调用相关函数增加日志
修复场景批量编辑失败的缺陷、为测试计划的定时任务和Jenkins调用相关函数增加日志

+++ backend/src/main/java/io/metersphere/api/service/ApiScenarioReferenceIdService.java
@@ -43,6 +43,9 @@ public class ApiScenarioReferenceIdService {
     }
 
     public void saveByApiScenario(ApiScenarioWithBLOBs scenario) {
+        if(scenario.getId() == null){
+            return;
+        }
         this.deleteByScenarioId(scenario.getId());
 
         long createTime = System.currentTimeMillis();
